+++ allennlp/training/metrics/boolean_accuracy.py
@@ -2,7 +2,9 @@ from typing import Optional
 
 from overrides import overrides
 import torch
+import torch.distributed as dist
 
+from allennlp.common.util import is_distributed
 from allennlp.training.metrics.metric import Metric
 
 
@@ -44,6 +46,7 @@ class BooleanAccuracy(Metric):
             A tensor of the same shape as `predictions`.
         """
         predictions, gold_labels, mask = self.detach_tensors(predictions, gold_labels, mask)
+        device = gold_labels.device
 
         # Some sanity checks.
         if gold_labels.size() != predictions.size():
@@ -84,6 +87,14 @@ class BooleanAccuracy(Metric):
         # where the entire prediction is masked (because they look "correct").
         self._correct_count += (correct * keep).sum()
         self._total_count += keep.sum()
+
+        if is_distributed():
+            _correct_count = torch.tensor(self._correct_count).to(device)
+            _total_count = torch.tensor(self._total_count).to(device)
+            dist.all_reduce(_correct_count, op=dist.ReduceOp.SUM)
+            dist.all_reduce(_total_count, op=dist.ReduceOp.SUM)
+            self._correct_count = _correct_count.item()
+            self._total_count = _total_count.item()
 
     def get_metric(self, reset: bool = False):
         """