+++ tests/training/metrics/entropy_test.py
@@ -1,7 +1,12 @@
 import torch
 from torch.testing import assert_allclose
 
-from allennlp.common.testing import AllenNlpTestCase, multi_device
+from allennlp.common.testing import (
+    AllenNlpTestCase,
+    multi_device,
+    global_distributed_metric,
+    run_distributed_test,
+)
 from allennlp.training.metrics import Entropy
 
 
@@ -15,18 +20,18 @@ class EntropyTest(AllenNlpTestCase):
             device=device,
         )
         metric(logits)
-        assert metric.get_metric() == 0.0
+        assert metric.get_metric()["entropy"] == 0.0
 
     @multi_device
     def test_entropy_for_uniform_distribution(self, device: str):
         metric = Entropy()
         logits = torch.tensor([[1, 1, 1, 1], [1, 1, 1, 1]], dtype=torch.float, device=device)
         metric(logits)
-        assert_allclose(metric.get_metric(), torch.tensor(1.38629436, device=device))
+        assert_allclose(metric.get_metric()["entropy"], torch.tensor(1.38629436, device=device))
         # actual values shouldn't effect uniform distribution:
         logits = torch.tensor([[2, 2, 2, 2], [2, 2, 2, 2]], dtype=torch.float, device=device)
         metric(logits)
-        assert_allclose(metric.get_metric(), torch.tensor(1.38629436, device=device))
+        assert_allclose(metric.get_metric()["entropy"], torch.tensor(1.38629436, device=device))
 
         metric.reset()
         assert metric._entropy == 0.0
@@ -41,4 +46,18 @@ class EntropyTest(AllenNlpTestCase):
         )
         mask = torch.tensor([False, True], device=device)
         metric(logits, mask)
-        assert metric.get_metric() == 0.0
+        assert metric.get_metric()["entropy"] == 0.0
+
+    def test_distributed_entropy(self):
+        logits = torch.tensor([[1, 1, 1, 1], [1, 1, 1, 1]], dtype=torch.float)
+        logits = [logits[0], logits[1]]
+        metric_kwargs = {"logits": logits}
+        desired_values = {"entropy": 1.38629436}
+        run_distributed_test(
+            [-1, -1],
+            global_distributed_metric,
+            Entropy(),
+            metric_kwargs,
+            desired_values,
+            exact=False,
+        )